Remove ad-hoc KLabel generation, simplify AMIT pass
"AutomaticModuleImportsTransformer" was doing 4 things,
only one was necessary and I've moved that to a new simplified
DeclareCellLabels pass

- adding two imports
This seem to be redundant with ones we now add earlier
- Adding declarations of KLabel constants for every constant mentioned anywhere
This was only necessary because three or four important things like isSet
were never declared anywhere in K, after adding them to the prelude this is
unecessary.
- Adding an explicit production subsorting every other sort into K
Removing this breaks no tests. If it turns out passes need to reason about
subsorting, it's probably better to special-case K.
- Declaring a CellLabel sort
This is only necessary in the generated Maude, so I've moved this
to happen very late.

git-svn-id: http://k-framework.googlecode.com/svn/trunk@9678 82a86efa-2583-11df-8ea0-49ba94de29ab

Former-commit-id: 5634b01b28b3a686297ce417ab5d1b2ab16c413b [formerly 3b5750cbe39c50409cb8410b7e3ae6105e7084ea]
Former-commit-id: f8ac00bed86a0527a948e0b97ac2dfe89896e079