Make DefinitionHelper fields non-static.
This commit finishes the job of making DefinitionHelper non-static,
by making the mutable fields non-static.

Only two places seem to be affected.
One is TermCons.toString which seems to be basically unused.

The other is Rewrite.getSort, which can be called with a null DefinitionHelper
inside of JavaClassesFactory, because the constructor for a KApp from an Element
uses sort information to decide whether to wrap the argument into a singleton KList,
and Rewrite.getSort uses information from the DefinitionHelper
(this situation occurs compiling logic.k for example).

We might fix this by making JavaClassesFactory do something simpler and arranging
the wrapping as a latter step in parsing, but for now I've left the logic alone
by instead giving JavaClassesFactory a DefinitionHelper, and adding some extra
methods for setting/unsetting it around construction jobs
(with a few asserts to try to complain about concurrent uses).

git-svn-id: http://k-framework.googlecode.com/svn/branches/v3.2@10038 82a86efa-2583-11df-8ea0-49ba94de29ab

Former-commit-id: 9b688a91f96f01d3cfbbb75fc90853c8cbc35b70 [formerly b0bd08cadc00b8e783c608aecede095fad340558]
Former-commit-id: 94b31e9d0a1310c7ec9311d08a0f830f9bf2051c