Initial Coq backend.
This adds --backend option coq
which generates Coq files defining the sorts and rules of a definition.

This first generates a file containing a parsed version of the rules of
the definition, and then runs a separate "kcoq" binary which generated
the actual Coq files by reading rule bodies from the generated file and
reading syntax from the actual .k files. The kcoq binaries are not
included in this commit.


Former-commit-id: 7d49766ad8e8b1f463d48edf951f38f9126a9287 [formerly 5ce1679024d0c1a555d0fa49975f9a6d26a31b1a]
Former-commit-id: eb33d402b713150a17ff6ecf12d45ef884e4d089