Declare isKResult with sort information
isKResult was previously declared as a KLabel constant,
now it is declared as a function in Bool.

Ths test output changes because a bare KLabel and a function-style
production pretty-print with different whitespace.
The modified regular expression explicitly ignores whitespace
rather than hardcoding one or the other, so it works with both versions.


Former-commit-id: 562c7b994f8937cb9002a2dd9b04e2e3ec4ddd5f [formerly caf4703bf8ea15e7d57e5df5c00e3aab9805a12e]
Former-commit-id: fdb4372067f1a1955513e5b1b6be56d19bdbb857