Let wait for kcoq process be interrupted
The best a Backend's run method can do with InterruptedException
is to be sure to set the thread's interrupted status before it
returns.
This patch immediately gives up on compiling if it gets interrupted
while waiting for the child process to finish, and returns immediately
from the backend after killing the child process.


Former-commit-id: 7def6709b7a8554b78eb496afbf614523ad338c5 [formerly f9415b1c3fa58de88800ca35d69658eb26e089bb]
Former-commit-id: a75424d42f3a73f525b3a72815ff4e9d58ea8794