Refactor isSubsort into syntactic and semantic check.
The previous Production.isSubsort method would consider
also productions with an explicit label to be subsort productions.

This commit renames isSubsort to isSyntacticSubsort, and
extracts the then-common pattern
  (prod.isSyntacticSubsort() && prog.getKLabel() == null)
to a new Production.isSubsort method.
A few other calls were changed from isSyntacticSubsort to isSubsort
when a local change could preserve the exact behaviour.


Former-commit-id: 2bf91ed1afc5092699ccc6b65b16d634d0dcc2f1 [formerly 92a9f77bdc54ef2f3f8bff4d5468424b78928f20]
Former-commit-id: 05e344d6402b8f46eb343b24d2f00f6cc69fdff5