Split semantic and syntactic subsort in Context
The subsorts member of Context was populated and used extensively
by the parser to represent the syntactic relation of a term of
one sort being legal in the concrete syntax where one of the other
is expected. In particular, this considers list elements subsorts
of the list type, and a production consisting of a single nonterminal
produced a subsort relationship whether or not it was labeled.

Unfortunately, other passes used this information from the context
as if it was the semantic subsorting relationship.
This commit splits the subsort information in Context into a syntactic
and semantic relation with independent accessors, changes
CollectSubsortsVisitor to populate both, and calls the appropriate
version in various places.


Former-commit-id: 09f8bd1d2ca4d7a2f7f73482c6605a59177ef020 [formerly feb23ce17ef4066374813cb03b8931b463b67953]
Former-commit-id: 8fe6eee0653ad516f1b11685dc747bdfc49ec22e