Adding #time function
#time should return the number of milliseconds since the epoch. However,
the return type is long, and long isn't supported in K as far as I know.
I made this commit, even though it's a mess, to get feedback from the
developers on how to properly do this. I was following [this pull
request](https://github.com/kframework/k/pull/967/files) which is why I
put it under IO but then I changed my mind in hooks.properties and made
time.k!


Former-commit-id: 5d6aab64e6338cc84b21866c3d71df4806dcd11c [formerly 009461fb4f1945ace59dce7940293ba7620ea770]
Former-commit-id: 160d00266dd8e3c994fd1f070dddae6404039be7