Move Coq backend to separate maven module
This commit also modifies test code in kernel because some
unit tests used on the CoqBackend, apparently just as a convenient
Backend instance to test injection.
The dummy TestBackend class is added as a replacement, and the test
setup is modified to register it under the name "test".

imp.labeled_rules was regenerated with the parent of this commit,
and is modifed only because of other changes in the K framework.
Extracting the Coq backend to a module did not change it further.
The Coq file was not tested or updated, because it's made by the
external kcoq program instead of the Java Coq Backend code.


Former-commit-id: 17b4de9f16faa0ae7d2498351283e058fcb96fd2 [formerly f0dd8b36fb376a7b313fa67b9a338b90ea370d18]
Former-commit-id: 08278e072b21221abe2c3dd9c56c88be8a079783