Unit test that a handwritten kore definition parses.
The kore file is based on a version of simple-untyped.k modified to be
self-contained (to compile and pass all tests with --no-prelude).
Rules are all written in labeled form, syntax declarations are
broken down into individual productions to follow the more basic
KORE outer syntax, and the configuration is encoded with syntax productions.
Configuration abstraction has not been applied.


Former-commit-id: b26a77031f1a6b796385af1a71476082ea57a507 [formerly 3748cbf0c6e790752adb9c25bd772fa66b75de7f]
Former-commit-id: 64b50970a56e76e17077a1345cc00c0f8dc2721d