Reduce heap size option in kserver.
With the previous large value the ktest run in mvn verify
would approach or exceed 8GB ram
(which is bad on my laptop that only has 8GB ram),
this change seems to keep it at least down to 5GB or so.
We suspect the GC was being too lazy when started with a huge
maximum heap size.
There also seem to be issues, not addressed by this commit,
with extra threads hanging around from threadpools that are not
properly cleaned up.


Former-commit-id: b537ac46ce8a827625e16a10dac7dcf6f749a658 [formerly d1fdbde34d870dbcae9114ef53613e136ea9f9c2]
Former-commit-id: eea8cd8a6cb39cb5507bb87d594118b6225c9575