Move developer instructions to repo root
After this commit, developer instructions for working from source are in
the repo root, user instructions meant for a release zip are in the
k-distribution directory, and the maven assembly file makes sure it's
still the user instructions that end up at the root of k directory in a
release zip.

This does not change the content of the files, besides adjusting
references to the developer documentation, and updating a copyright header.


Former-commit-id: 7474defe4597de3c8dafa6abb08aec36ba6f06a2 [formerly 2f13cb94279d101b3ff70e89c022be322c9a374e]
Former-commit-id: 7d5363e03216327bb2bea4025af0fbb6c07861b8