Refactor parseString to return K.
All callers immediately applied TreeNodesToKORE.apply,
so it's convenient to add it into parseString.

Also, a recent change to AmbFilter makes the result somewhat
unpredictable as a Term by picking an arbitrary resolution of
any ambiguity where the unequal Terms become equivalent
after applying TreeNodesToKORE.apply, so it's safer not
to expose the Term-returning version.


Former-commit-id: 84082ec64badd979d48457b71e2278f8cd0a0870 [formerly 1781109e9322489b906e4aa03f490a74bff958b5]
Former-commit-id: 0dab790697fc098795be324c53717c5e1611e1b6