Support cell rewrites.
The SortCells pass does not recurse to find cell rewrites hidden under
labels such as "#cells", and was thus incorrectly computing the sets of
neighboring cells of some cell fragment variables.
This commit fixes that. It also simplifies the logic by adding
a small pass to lift any rewrite inside a non-leaf cell to occur
as an immediate child of the parent cell.


Former-commit-id: db59d59c6f1f2c2a6d687753eb4256575aca3590 [formerly b713b0718871747af6e88c35f78848d4a319c4b9]
Former-commit-id: 9592a277f813aeae9acec00794df1184c08c7847