Parse parent cell contents as "Cells" in rule gramamr.
Instead of parsing as K, parse the contents of non-leaf cells
as "Bag" (which is the sort of unsorted and uncompleted
collections of cells).
This allows the loose productions generated for cells
(allowing dots, and unsorted contents of parent cells) to be
placed in the proper cell sorts and only subsorted into Cell,
without introducing parsing ambiguities between
  Bag ::= Bag Bag
and the more specific productions for particular multiplicity * cells.
Bag includes KBott, so this doesn't exclude rewrites or label applications
 - and terms with least sort K would already break the cell sorting pass.


Former-commit-id: ec54eb9bde38d61d5052ddb3f2f1ae892550d1d6 [formerly 6ecf4c1ff3e96c40631bc745a66c6088910511ec]
Former-commit-id: 6c20c2c73608e09e0dcad4cc33c18466134b1028