Filter argument sorts
AddEmptyList now computes the sort of terms in labeled syntax more carefully.
The second argument of the #KApply is divided into arguments by
splitting on #KList applications.
If none have sort KList, it is assumed to be a list of concrete
arguments, and the productions for the klabel are filtered down
to those consistent with the arguments, before returning
the least sort among the remaining results.


Former-commit-id: fb03b29f3437d2a8eee400ff9bb5642655f9a437 [formerly b68f84991200e0bac90c2205ad2691c9bf067a90]
Former-commit-id: d72e8aba784a9b7adb47ea192b1474a74a5cc147