Fix ocamlopt kompile and provide fakelibs for c-semantics build
Two issues:
1) The apparently-useless copy of execution_pgm.ml removed in an earlier
commit was actually the only code creating the temp directory,
so attempting to redirect process output to files in the temp directory
was broken.

For this, FileUtil was changed so resolveTemp creates the temp directory
if needed.

2) After fixing the above, ocamlopt was failing to find -lgmp and so on
while trying to link the -output-complete-obj output, because the
fake lib*.a libraries were not found. The Java code assumed they
were included in k-distribution/include/ocaml/fakelibs, but they
had not been committed there, and the c-semantics-plugin makefile was
creating a fresh copy in the build recipes.

For this, the fakelibs were committed in the expected place
(with a new .gitignore file to override an earlier suppression of *.a),
and also the c-semantics-plugin makefiles were updated to use them.