KILtoSMTLib: Improved translation of constraint and implication queries (#212)
* KILtoSMTLib and other classes: Refactorings and comments for clarity.

* KILtoSMTLib: 
1 Improved translation of constraint and implication queries.
A term that has no smtlib attribute will be abstracted to a fresh anonymous variable. Previously the whole LHS/RHS of the equality that contained the term was abstracted to a variable. This lead to significant information loss making most Z3 queries useless.
This translation was also enabled for implication RHS now. It is proved sound.

2. Making Z3 queries a bit prettier for easy debug.

* KILtoSMTLib: "Middle ground" Solution for translation to Z3. Do not allow substitution of untranslatable terms with fresh variables, if these vars appear only in the RHS of the implication. Allow them if they appear both in LHS and RHS. Performance improvement.