Java Backend KEVM customization: Detecting final state, when first item in <k> is #halt and there is nothing left to execute. If this state is reached and target term is not matched, execution is halted.
Useful in kprove, to prevent further symbolic execution and needless branching that would happen if spec would continue to evaluate past intended final state.

If specification does not use #halt in target term or is non-KEVM, this customization has no effect.
Temporary. Was replaced by more general and semantics-agnostic `--halt-cells`, that is about to be PRed.