use only minimal productions when creating ambiguities during unparsing
This solves an OutOfMemoryError when unparsing large associative lists
of polymorphic productions, which were exploding in the size of the tree
during unparsing. We should never need to choose something that is not
minimal with respect to the overload relation because a minimal
production can always go anywhere a production would go that is greater
than it.