Java Backend: function cache fix for impure/freshVar (#354)
Added a field `JavaSymbolicObject.isCacheable`, computed recursively on the AST tree. It is only computed on fully evaluated terms. The value depends not only on the present term bot also on the source term from which present term was computed through function evaluation.
A result term is not cacheable if either:
1. Source term KLabel is `[impure]`.
2. A fresh variable is created during evaluation from source term to present term.
3. Any sub-term is not cacheable.

The entry `source -> result` is added to the cache only for cacheable results. But the entry `result -> result` is added for any term, to distinguish terms that are already evaluated from unevaluated terms.

Extra:
* regression-new/symbolic-int-function: regression in generation of fresh integers

* Refactoring: replaced usages of KItem.evaluateFunction() with KItem.resolveFunctionAndAnywhere() to remove a fragile duplicate.