use default java serialization again
FST doesn't work with Java 16, nor is it likely that it can be easily
fixed. The same problems apply to kryo. We might be able to make a
serializationn library work if we specify how to serialize each type of
object that we serialize, but that would be a lot of work. I'm not
saying we shouldn't do that, but I wanted to present this alternate
solution so we can consider the pros and cons and do some performance
measurements.