K-tutorial: make several edits (#2087)
* K-tutorial: Lesson 11: rename isStmt to isStatement

It appears that `isStmt` is a function defined as a part of the BOOL
library. Change the name of this function to that the those who are
working through the tutorial do not write a function that causes
confilcting definitions for `isStmt`.

Signed-off-by: Erik Kaneda <erik.kaneda@runtimeverification.com>

* K-tutorial: clarify a few phrases

The rewording is purely for clarification and is not intended to change
the over all meaning of the surrounding text.

Signed-off-by: Erik Kaneda <erik.kaneda@runtimeverification.com>

Co-authored-by: Everett Hildenbrandt <everett.hildenbrandt@gmail.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>