Allow LESSON-05-D tutorial file to compile without flags (#2159)
* Allow LESSON-05-D to compile without flags

All previous lessons compile without the use of the --main-module flag.
While this flag does get explained further on in this tutorial, by
changing the wording here it avoids potential confusion on why the
command no longer works.

* Update tutorial test command list

The changes to the tutorial markdown made in the previous commit need to
be reflected in the command list.