pyk tests: remove -I from kompile add -I to kprove
This test case utilized -I in the kompile command in order to pass the
parameter of this option to kprove. This was done through serializing
the `kompile` object during the invocation of the `kompile` command and
de-serializing the `kompile` object to access the `outerParsingOption`
during the invocation of the `kprove` command. Since the -I option has
been added to `kprove`, drop this command from the `kompile` command and
move it to the `kprove` command.