kprove-markdown: use --md-selector for testing
This test adds additional lines in a test file to test the new --md-selector
option for kprove. Before this feature was added, the default behavior was for
kprove to use k as the selector. In order to test that the --md-selector
option is functional, text has been added to the test case so that it will
cause a failure on the default md-selector.