Capture git --describe using antrun
When the project is compiled, the output of git describe is captured as
a Maven property; this property is then stored in the packaged .jar
file's manifest (similarly to existing properties like the build
timestamp).

This implementation doesn't rely on an external plugin, but does require
a git executable (other antrun tasks for copyright checks have similar
requirements). It will fail silently if the package is built from a
source distribution with no .git directory such as a tarball, and the
revision will not be printed.