Include git describe in output of kompile --version (#2184)
* Capture git --describe using antrun

When the project is compiled, the output of git describe is captured as
a Maven property; this property is then stored in the packaged .jar
file's manifest (similarly to existing properties like the build
timestamp).

This implementation doesn't rely on an external plugin, but does require
a git executable (other antrun tasks for copyright checks have similar
requirements). It will fail silently if the package is built from a
source distribution with no .git directory such as a tarball, and the
revision will not be printed.

* Fix silent failure behaviour

If an error property is not specified, stderr gets redirected to stdout
and the resulting 'fatal: ...' message ends up as the revision ID. This
commit fixes this by explicitly setting an error property, even though
it is not referenced later.

* Include only one version string