Fix version flag tests broken by git describe (#2187)
This test previously expected to see the exact version string supplied
by the package version file when --version is passed to the K tools. In
PR #2184, the output of --version was changed to print the (more
precise) git revision rather than the package version.

This commit fixes the test by allowing _either_ the output of git
describe or the package version to appear in the output of --version.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>