Implement -I, --md-selector, and --no-prelude in kprove, kprovex, keq, and kbmc (#2172)
* Extract -d from OuterParsingOptions

This extraction prepares OuterParsingOptions class to be included in the
following tools: kprove, kprovex, keq, and kmbc.

* KProve, KEq, KBMC: add OuterParsing options

Adding this object enables the `-I`, `--md-selector`, and `--no-prelude`
options to these tools. By doing so, each tool can configure their own
OuterParsing options rather than getting this information by de-serializing
the kompile object from the cache left by a previous invocation of the
`kompile` command.

* pyk tests: remove -I from kompile add -I to kprove

This test case utilized -I in the kompile command in order to pass the
parameter of this option to kprove. This was done through serializing
the `kompile` object during the invocation of the `kompile` command and
de-serializing the `kompile` object to access the `outerParsingOption`
during the invocation of the `kprove` command. Since the -I option has
been added to `kprove`, drop this command from the `kompile` command and
move it to the `kprove` command.

* add code review suggestions

* Compiled definition: add outerParsingOptions

* KompileFrontEnd: use injected OuterParsingOptions

* kprove-markdown: use --md-selector for testing

This test adds additional lines in a test file to test the new --md-selector
option for kprove. Before this feature was added, the default behavior was for
kprove to use k as the selector. In order to test that the --md-selector
option is functional, text has been added to the test case so that it will
cause a failure on the default md-selector.

* Kprove: use KproveOptions.outerparsing to initialize kompiledDefinition

* KompileOptions: make OuterparsingOptions transient

Co-authored-by: rv-jenkins <admin@runtimeverification.com>