Sort output from kdep tests (#2202)
* Sort output from kdep tests

This should fix the flakiness issue where these tests would
intermittently fail on some CI runs but not others.

* Revert "Sort output from kdep tests"

This reverts commit daed04273d70196d6ddf4647fe912aacb787d9bb.

* Sort in Java code, test style fixes

This commit moves the sorting of dependency files into the Java
frontend, rather than applying a sort to the test output files.

Additionally, it fixes stylistic problems with the testing code.

* Update sort order and results