Add explicit to_int cast to ^Int SMT hook
This was recommended as a workaround for #2154 and #1957 by the Z3
maintainers (https://github.com/Z3Prover/z3/issues/5243).

The original issue arises because Z3 cannot assume that the exponent is
non-negative, and so must infer that the result could potentially have
type Real after evaluating the exponentiation. By adding the to_int
cast, we opt in to integer semantics everywhere ^Int is used.

Additionally, this commit adds a simplification rule for integer
exponentiation that sends negative exponents to zero directly from K.