Add #kompiledDirectory hook to K-REFLECTION (#2209)
* Add hooked function to K-REFLECTION module

This exposes the kompiledDir function provided by the LLVM backend to
the K frontend.

* Test #kompiledDir hook

These tests cover two cases - one where the definition directory is the
current working directory, and one where it is a nested subdirectory.
The test itself just checks whether the function `#kompiledDir`
evaluates to the correct value after the current working directory has
been substituted out.

* Update test results affected by reordering

The changes to domains.md seem to have caused the outputs of these tests
to become reordered - this commit updates them to the new ordering as
produced by the update-results rule.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>