Expose floating-point truncation hook (#2218)
This function mimics the style of the floor and ceiling operations, nd
is a direct passthrough to the MPFR implementation of floating-point
truncation.

There don't seem to be any integration tests for ceil / floor at the
moment; they're only unit-tested in the backend. If it would be useful
to have tests for all these functions, I can implement them.