pl-tutorial: Add token label to `Id` token sorts
These files contained token sorts called `Id`. The java backend seems
to allow token sorts to be defined with the `token` label or as strings.
The haskell and llvm backends only allow token sorts to have productions
that are tagged with the `token` label. Since the java backend will be
deprecated, update the sort productions to include the `token` label
so that it will work on the newer backends as well as the java backend.