kompile: add check for token sort productions (#2241)
* kompile: add check for token sort productions

Token sorts should only contain productions that have `token`,
`bracket`, or `function` labels or the RHS of the production is a macro
term. Any other label (including subsorts) are illegal.

* invalidConstructor check: update test results

* regression-new: add tests for token check

* kompile: do not run token checks when running java backend

The java backend is more forgiving when it comes to token sort
declarations. Avoid this check when kompile is running with the java
backend.

* pl-tutorial: add token labels to files kompiled using llvm backend

* CheckTokens: move conditional to structuralChecks

CheckTokens class does not need to evaluate this logic for every sentence.

* Add comment.

Co-authored-by: Radu Mereuta <headness13@gmail.com>