kompile: check rules with smt-lemma labels (#2248)
* kompile: check rules with smt-lemma labels

Rules marked with the smt-lemma label must be written in a way where
all terms (aside from tokens) are labeled with `smtlib` or `smt-hook`
attributes. Without these labels, it's not possible to encode
these rules for Z3.

* add test for smt-lemma checks

* Address code review feedback

Remove the override method for KRewrite. The same logic is implemented
in the super class.

Fix logic that searches productions for labels.

* remove unnecessary check

* clarify error message

* update test results