kompile: emit error on macro terms that failed to expand
We are in the process of enabling macros on productions and disallowing
macros on rules. The ability for users to use macro productions also
necessitates the need for them to correctly write rules that translate
the macro term. This means that it is possible for a macro in a rule to
not be expanded. Add a check to ensure that no macros exist after
expansion.