Check for existential variables in requires clause (#2336)
* Check for existential variables in requires clause

The `requires` clause is a part of the LHS of the rule. Since existential
variables are not allowed on LHS of the rule, these variables are not
allowed in the clause. Add a check to catch rules like this.

* Update existing test with existiential variables in requires clause

This is a part of a change set that fixes kompile so that it can detect
existential variables in `requires` clauses. Existential variables are
not allowed on LHS of rules/claims and the `requires` clause is a part
of the LHS. Update the existing test to add this case.

* Remove existential variables from test cases

These tests do not require the existential `?`. Remove them.

* update test case result

* Revert "update test case result" and remove additional `?`

Remove an additional `?` that was causing the test output to change.
This means that we can revert commit 60d41d5bdfeb150bca04f750ca598ce111cfc334.

* Add _ to unused variable names