Allow macros on hooked sorts (#2489)
The previous implementation of macros on productions forgot to account for
macros on hooked sorts. This led kompile to assume that macro productions on
hooked sorts implied that the productions were constructors. Code like this

```
syntax Int ::= "LABEL" [macro]
rule LABEL => 4 // or some other number
```
would result in errors like this:

```
$ kompile test.k
[Error] Compiler: Cannot add new constructors to hooked sort Int
```

However, `"LABEL"` in this case is not a constructor but a term that
would be translated to `4` during compilation so this error is not
applicable to productions that are macros. This change add changes to
the CheckAtt class to allow macros on hooked sorts.

Co-authored-by: rv-jenkins <admin@runtimeverification.com>