Print source lines in error messages (#2517)
* Print source lines in error messages

At the time of writing, error messages display location information like
so:
```
[Error] Compiler: Found existential variable not supported by concrete backend.
        Source(/home/erik/error/test.k)
        Location(11,39,11,41)
```

Sometimes these messages can be difficult to decipher due to the lack of
information from the actual source code. Add this information in the
error message so that the errors look like this:

```
[Error] Compiler: Found existential variable not supported by concrete backend.
        Source(/home/erik/error/test.k)
        Location(11,39,11,41)
        |    rule <k> a => b ... </k> requires ?_ ==Int _
        |                                      ^~
```

For longer lines, error messages look like this:

```
[Error] Compiler: Claims are not allowed in the definition.
        Source(/home/erik/error/test.k)
        Location(15,11,17,13)
           .          v
        15 |    claim _
        16 |          =>
        17 |           .
           .         ~~^
```

If error messages span more than 3 lines, the lines between the first
and last line are replaced with a single line containing `...` to avoid
taking up too much space on the terminal.

* regression-new/checks: update expected test output

* regression-new: update results other than checks

Co-authored-by: rv-jenkins <admin@runtimeverification.com>