Add a check to detect duplicate k cell declarations (#2827)
* Check for duplicated K cell declarations

* Add kompile tests for duplicate cells check

* Use visitor to inspect cell names instead

* update test results

* Update k-distribution/tests/regression-new/checks/duplicateKCells.k

Co-authored-by: Radu Mereuta <headness13@gmail.com>

* Update k-distribution/tests/regression-new/checks/duplicateKCells.k

Co-authored-by: Radu Mereuta <headness13@gmail.com>

* Applying Radu's feedback

Radu's suggestion is much more concise. The solution to emit an error on
all K cell declarations has been retained because it allows the user to
see all locations of all k cell declarations.

* Add additional case to test and update results

* Remove unused code

This code was used in a previous solution and no longer needed. Remove
it.

* Throw exception if multiple K cells are found

There are assertions that cause a crash in later stages of kompile so
abort compilation when multiple K cells are found.

* update test results

Co-authored-by: Radu Mereuta <headness13@gmail.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>