Add --keep-cells to pyk print (https://github.com/runtimeverification/pyk/pull/42)
This PR adds the `--keep-cells` functionality to `pyk print`.

# Meaning
The `--keep-cells` option will make `pyk print` to keep only indicated
cells *and "skeletons" of their parents* in the input configuration, the
rest is substituted with `...` and minimized accordingly.

# Usage

Usage is the same as `--omit-labels`. Let's say we have the following
configuration:

```
<kevm>
<gas> 0 </gas>
<pc> 0 </pc>
<nonce> 0 </nonce>
</kevm>
```

Then running `pyk print {PATH_TO_DEF} {PATH_TO_FILE} --keep-cells
"<pc>,<nonce>"` will roughly return

```
<kevm>
<pc> PC </pc>
<nonce> NONCE </nonce>
...
</kevm>
#And {POSSIBLY_PATH_CONDITION}
```
(with `PC` and `NONCE` depending on the current contents of the cells).

*Note*: `--keep-cells` cannot be used together with `--omit-labels`.
Setting both will make the CLI throw an exception.

# Documentation

A description was added to `pyk print -h`.

# Quirks

Currently, only cells with primitive values are supported! Selecting
cells containing other cells will return an empty term. Consider that
this is actually correct: in this case we want to keep cells, which
contain cells that we do not want to keep => the contents of the
selected cell get rewritten to `...` => even the selected parent cell is
extracted.
*Workaround*: choose the cells with primitive values you actually want
to see, then their parent is included as well. There's no option to
select multiple cells simply with their parent cell. If there's a large
cell with several children cells that you want to track, consider using
`--omit-labels` instead.

Co-authored-by: devops <devops@runtimeverification.com>