Two Nix fixes; update nixpkgs (#60)
Currently, `poetry2nix` tries to build development dependencies of the
Python project. That (1) is totally unnecessary, and (2) does not work
with some newer versions of nixpkgs. This PR fixes the problem, which
allows `nixpkgs` to be updated again. Furthermore, project that depend
on this flake do not have to transitively depend on the dev
dependencies.

The fix consists of two parts:
* In `pyproject.toml`, the line `[tool.poetry.dev-dependencies]` is
replaced by `[tool.poetry.groups.dev.dependencies]`, which is the newer
way of saying the same thing. Poetry2nix seems to support only the newer
syntax.
* Since `poetry2nix` by default tries to resolve also the dev
dependencies, we say (by `checkGroups = [];`) that it should not do
that.

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Samuel Balco <goodlyrottenapple@gmail.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>