Implement `execute` method via RPC server in `KProve` class (https://github.com/runtimeverification/pyk/pull/81)
This PR takes some steps towards:
https://github.com/runtimeverification/hs-backend-booster/issues/3

I made an initial attempt to have the new KEVM rewriter use kore-rpc to
query the old backend (as a replacement for textual `kprove` interface),
and ran into several limitations in our Kast <-> Kore conversions and
had to figure out the logic for wrapping the RPC server to make the
calls.

- Adds the `depth` option directly to the various methods in `KProve`
class.
- Handles Kast <-> Kore conversion for variables, with several tests
added of examples that were run into while trying to handle KEVM.
- The munger/unmunger had to be modified to deal with KVariable as well
as KLabel.
- The `_add_sort_injections` pass is made to handle the case where the
input/output sorts are the same, so that we don't need to check that
they are different at the call sites.
- The built-in data-structures conversion routine will print a better
message about which specific subterm can't be converted between the two.
- Adds a method `KProve.execute`, which wraps the RPC server in the
following way (tests are added):
- The first time you use `KProve.execute`, a server and client are
started up, and the `execute` endpoint of the server is called.
  - The server is kept around for subsequent calls.
- It has roughly the same interface as `get_claim_basic_block` so it can
be plugged into downstream libraries. We may need to change this, but if
not then we may be able to simplify the `execute` endpoint.

So the server will only be started if it's needed, but won't need to be
restarted for subsequent queries. I'm hoping this gives us better
overhead with the backend communication.

Co-authored-by: devops <devops@runtimeverification.com>