Improvements in Kast -> Kore conversions (https://github.com/runtimeverification/pyk/pull/93)
This PR builds up the functionality needed for KEVM to do KCFG
exploration using RPC endpoints, instead of textual interface.

- Add Kast <-> Kore updates:
  - Conversion for `#Ceil` is added (and test).
- Conversion `kast_to_kore` is modified to regularize sorts annotating
variables before doing the conversion (and test).
- Keeps `build_rule` from renaming variables to themselves, so that it
does not throw away any sort information that some of the variables may
have (and test).
- Adds `KDefinition.sort_vars : KInner -> KInner` and
`KDefinition.sort_vars_subst : KInner -> Subst`. These methods are used
to normalize the types of variables showing up in terms to all be the
same sort (needed for converting Kast -> Kore, so that all converted
variables have the same sort).
- This uses new helper `var_occurances : KInner -> Dict[str,
List[KInner]]`, which computes occurances of variables showing up in
terms, but keeps the lists of occurances. This method is also used in
`free_vars` and `count_vars`.
- `KCFG.create_node` is modified to not strip generated cells upon
inserting new nodes. This is needed for being able to take the output
and put it back into the backend. Now `build_rule` removes the cells
directly before constructing the rule for the frontend (hold onto the
cells as long as possible).
- A now unused function (checked here, KEVM, and KSummarize)
`kastManip.mark_useless_variables` is removed. It's logic was inlined
into `build_rule` a while ago.

Co-authored-by: devops <devops@runtimeverification.com>