Updates to `KProve.execute` endpoint (https://github.com/runtimeverification/pyk/pull/100)
This makes the changes needed to get the python and kore-rpc rewriters
to agree on executions out to depth 60 on KEVM.

- Adds method `State.kore` for constructing the overall return-pattern
of `kore-rpc` execute endpoint.
- `KProve.execute` now returns more information, returning the next
state and list of successors if present.
- Adds several helpers to `KDefinition`, including `alias_rules`,
`macro_rules`, and `semantic_rules`.
- Memoize computation of `KDefinition.init_config` and
`KDefinition.empty_config`, they are the biggest cost center for using
`KRun.run*`.

Co-authored-by: devops <devops@runtimeverification.com>