Allow larger terms to be passed to `KRun.run_kore_config` (#104)
Currently, `KRun.run_kore_config` relies on passing configuration
variables on the CLI as arguments. But this is bad for large arguments.

This changes `KRun.run_kore_config` to use `KRun.run_kore_term`. The
logic it uses is exactly teh same as the current `krun` bash script will
use, but avoids having to pass the arguments on the command-line.
Specifically:

- The configuration variables are packed into a `Map` where the keys are
`KConfigVar` and the values are terms of the appropriate sort (call this
`CONFIG_VAR_MAP`).
- The initial configuration `initGenerateTopCell(CONFIG_VAR_MAP)` is
constructed, and this is what is passed to `KRun.run_kore_term`.
- Unrelated minor change: The `kompile(...)` helper defaults to emitting
json definition, because without it the pyk library is pretty useless.
Tests are updated accordingly.

Co-authored-by: devops <devops@runtimeverification.com>