Steps towards RPC prover (#112)
These changes are what I needed to have a first working version of
KProve implemented fully here:
https://github.com/runtimeverification/evm-semantics/pull/1443. In
particular:

- Some KAst <-> Kore problems are addressed:
- If the same variable shows up multiple times in
`KDefinition.sort_vars`, but with different sorts, instead of just
giving up we'll check if one sort is a subsort of the other. If so,
we'll return the smaller sort. Test added to `test_kast_to_kore`.
- The conversions for `Exists` are added. Test added to
`test_bidirectional`.
- `KProve.execute` method now returns `CTerm` instead of `KInner`, as it
shouldn't be able to return `mlTop()` or `mlBottom()` anyway.
- `KCFG.create_cover` now allows you to specify a witness manually
instead of trying to compute it for you. (De)Serialization requires that
we can do `Subst.{from_dict,to_dict}`, so that functionality is added
too.
- `KProve.implies` method is added as a wrapper around the `implies` RPC
endpoint. Tests are added that make sure it's giving back sane results.
By default, we much existentially bind variables in the consequent that
are not present in the antecedent to get anything useful out of the
endpoint. Tests added to `test_implies` in `test_kprove`.

Co-authored-by: devops <devops@runtimeverification.com>