Test of execute endpoint with cell map (https://github.com/runtimeverification/pyk/pull/142)
There was a bug in how KEVM was handling cell maps, and I could not
figure it out. So I made a minimal example to reproduce the bug, and the
last commit here demonstrates the fix.

So somehow the kore terms being generated made it past validation in the
backend (even without the wrapper mapitem labels), but this last commit
shows how to correctly build the terms so they can execute properly.

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Samuel Balco <goodlyrottenapple@gmail.com>
Co-authored-by: Tamas Toth <tamas.toth@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>