Better KCFG visualization (https://github.com/runtimeverification/pyk/pull/152)
This PR makes some improvements to the KCFG visualization, after
discussions with the textualize developers in [their
discord](https://discord.gg/qPJkUrxc).

- The `KCFG.pretty` is factored into functionality
`KCFG.pretty_segments`, which returns each _chunk_ of lines separately
with an identifier attached.
- The `KCFG.pretty` outputs more compact textual representation, which
is a bit more visually appealing.
- The `KCFGViewer` has better CSS styling/presentation.
- The `KCFGViewer` is more modular, instead of being one monolithic
application.
- You can now click on nodes in the KCFG viewer, and it will jump you
straight to that node.

See
https://drive.google.com/file/d/1ThtFn8hcTc3P5lsviz4DR2R6mZyJjCxc/view
for a video demo of new functionality.

Co-authored-by: devops <devops@runtimeverification.com>