Add initial bug-report generation functionality (https://github.com/runtimeverification/pyk/pull/99)
I've tested this with KEVM, and it produces bug reports for `krun` and
for `kore-rpc` that are reproducible _on my machine_. The `kore-rpc` bug
reports should be reproducible across machines, but the `krun` ones may
not be (machine specific binaries etc...).

This PR:

- Introduces the `BugReport` class.
  - This class tracks a tarball to attach bug reports to.
- It exposes facilities for adding files/commands to the bug report, so
that if bug reporting is turned on, it can collect those files/commands
directly.
- Threads the new `BugReport` functionality into `KRun` class, via the
`_krun` method, which means that it's now ready for any calls into
`KRun.run*`.
- Threads the new `BugReport` functionality into `KoreClient` and
`KoreServer` class, and into the `KProve.execute` and `KProve.simplify`
endpoints that use the RPC server.

Co-authored-by: devops <devops@runtimeverification.com>