Expect non-zero return code from `krun` (#160)
I have a use-case for the C semantics / FRET project where I'd like to
be able to use `krun` via Pyk to take a few steps from an initial
configuration. Our semantics initializes the return-code cell with a
non-zero value so that we can identify when rewriting stops before a
program has completed execution; this causes `krun` to raise an
exception when run via Pyk.

This PR adds an extra option to the `KRun` run methods that allows a
specific return code to be expected in such situations, as well as an
integration test-case for this.

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: Tamás Tóth <tothtamas28@users.noreply.github.com>