Allow overriding unparser for KPrint/KRun/KProve (https://github.com/runtimeverification/pyk/pull/185)
https://github.com/runtimeverification/pyk/pull/183 changes the way that
we cache properties in the `KPrint` class (to use more standard Python
features). But this breaks some functionality in KEVM where we were
manually overriding the unparsing definition and unparsing symbol table.

Here is a proposed solution:

- Thread optional arguments `definition` and `patch_symbol_table` into
each constructor.
- If these arguments are present, ignore the definition on disk and use
these ones.

Here is where I adapt KEVM to this scheme:
https://github.com/runtimeverification/evm-semantics/pull/1546

Some alternatives (perhaps better?)

- Instead of allowing overriding the entire `KDefinition`, allow
specifying `extra_modules` that should be included into the one read on
disk. This should be compatible with all the uses in KEVM.
- Instead of allowing to pass in `patch_symbol_table` as an argument,
make it a method with default implementation the identity. Then
subclases can override it as needed. This may be cleaner code-wise.

---------

Co-authored-by: devops <devops@runtimeverification.com>