Allow specifying depth in KCFGExplore.step, prune bottom branches (#190)
This adds a parameter `depth` to `KCFGExplore.step` so that users have
better control over how it expands a given edge.

- `KCFGExplore.step` takes a `depth` parameter which allows taking more
than one step from a given node in the KCFG.
- Changes `KCFG.section_edge` to reuse `KCFG.step` instead of
re-implementing some of the same logic.
- Checks in `KCFG.step` that edges which are being stepped on do not
have a condition.
- Adds debug logging for for `_kast_to_kore`.
- Adds a step to `KCFGExplore.execute` endpoint to check that returned
branches are actually feasible.
- Adds a test of `cterm_implies` where the backend was returning a bad
result, previously had a workaround on this branch:
https://github.com/runtimeverification/pyk/pull/184
- Adds an initial attempt to test that bottom branches are not being
pruned properly, but fails. Sets up the testing harness and adds new
test anyway.

---------

Co-authored-by: devops <devops@runtimeverification.com>