Fix minor bugs in unparsing attributes and add tests (#226)
I tried unparsing a definition read in from `compiled.json`, and noticed
taht pyk actually crashed on this. It's because the `Location` attribute
is stored as:

```
...,"org.kframework.attributes.Location":[2135,3,2135,20],...
```

And our unparser was assuming that the values in attributes were
strings, not more complex values. So this PR:

- Makes it so we can unparse anything that has string represenattion as
values in attributes.
- Adds a test that a definition read in from the frontend is
unparseable.
- Adds tests for unparsing certain special frontend attributes
correctly.

---------

Co-authored-by: devops <devops@runtimeverification.com>