Allow for custom KCFG views in TUI (#227)
This PR makes some advances to the KCFG visualization. In particular:

- A footer is added which displays the keypresses available to the user
(and acts as clickable buttons).
- A new class `NodeView` is factored out for controlling the right-pane
views (of term and constraint).
- Better information displayed in `Info` bar.
- Ability to toggle whether term minimization is turned on or not.
- Ability to supply a custom semantics specific view, which displays
under the term/constraint views.
- Ability to toggle `term`, `constraint`, and `custom` views on and off.
- Fixes a small bug in `kompile.py` where an f-string didn't have an `f`
in front of it, and had a bad value name.

Here is a video demo of the functionality implemented on KEVM for
displaying Solidity source maps:
https://drive.google.com/file/d/1GJDjlxQKeZLI1WKwiV2lM26ZP4TGRqwi/view?usp=share_link

---------

Co-authored-by: devops <devops@runtimeverification.com>
Co-authored-by: rv-jenkins <admin@runtimeverification.com>